OSX gcc, cast enum to int to be used with kdbgstream
    Some how harald's kdebug change removed some overloads which
    where used by the anon union/enum and this enum...

CCMAIL: harry@kdevelop.org

svn path=/trunk/KDE/kdebase/khelpcenter/; revision=608915

+++ formatter.cpp
@@ -99,7 +99,7 @@ bool Formatter::readTemplates()
         break;
       default:
         kError() << "Formatter::readTemplates(): Illegal state: "
-                  << state << endl;
+                  << static_cast<int>(state) << endl;
         break;
     }
   }
